split up contents into subfiles
