+++ src/gui/mod.rs
@@ -1,1 +1,5 @@
 pub mod connection;
+pub mod entry;
+pub mod gui_message;
+pub mod state;
+pub mod temp_message;
