+++ src/gui/state/mod.rs
@@ -1,0 +1,4 @@
+pub mod entry;
+pub mod new_entry;
+pub mod password;
+pub mod vault;
